Almost working. Top half works.
git-svn-id: svn://10.0.50.55/legends/trunk@1255 67207eb2-9daa-044d-9a34-b8567af74898

+++ code/src/GEL2/Grid/SubGrid2D_Raycast.h
@@ -118,7 +118,7 @@ inline void TraceShadowGrid( const SubGrid2D<u8>& Src, Grid2D<u8>& Dest, const i
 			for ( szt y = 0; y < Dest.Height(); y++ ) {	
 				for ( szt x = 0; x < Dest.Width(); x++ ) {
 					if ( Dest(x,y) == 0 ) {
-						Vector2D PointToStart(x1-x,y1-y);
+						Vector2D PointToStart( x1-(int)x, y1-(int)y );
 						
 						if ( (dot(PointToStart, RayATangent) < Real::Zero) && (dot(PointToStart,RayBTangent) < Real::Zero) ) {					
 							Dest(x,y) = 2;
